添加短信转移记录

+++ SMSStattion/src/com/unityprima/smsstattion/utils/DefaultSettingInfo.java
@@ -4,6 +4,6 @@ package com.unityprima.smsstattion.utils;
 public class DefaultSettingInfo {
 	private DefaultSettingInfo(){}
 	
-	public final static int DEFAULT_LOOP_CYCLE = 15;
+	public final static float DEFAULT_LOOP_CYCLE = 15;
 	public final static int DEFAULT_WARNING_POWER_THRESHOLD = 20;
 }
